Remove v1 from ecctool
