+++ ecchronos-binary/src/bin/ecctool.py
@@ -27,24 +27,22 @@ from argparse import ArgumentParser
 from io import open
 
 try:
-    from ecchronoslib import rest, table_printer, table_printer_v2
+    from ecchronoslib import rest, table_printer
 except ImportError:
     SCRIPT_DIR = os.path.dirname(os.path.realpath(__file__))
     LIB_DIR = os.path.join(SCRIPT_DIR, "..", "pylib")
     sys.path.append(LIB_DIR)
-    from ecchronoslib import rest, table_printer, table_printer_v2
+    from ecchronoslib import rest, table_printer
 
 DEFAULT_PID_FILE = "ecc.pid"
 SPRINGBOOT_MAIN_CLASS = "com.ericsson.bss.cassandra.ecchronos.application.spring.SpringBooter"
 
+
 def parse_arguments():
     parser = ArgumentParser(description="ecChronos utility command")
     sub_parsers = parser.add_subparsers(dest="subcommand")
-    add_repair_status_subcommand(sub_parsers)
     add_repairs_subcommand(sub_parsers)
     add_schedules_subcommand(sub_parsers)
-    add_repair_config_subcommand(sub_parsers)
-    add_trigger_repair_subcommand(sub_parsers)
     add_run_repair_subcommand(sub_parsers)
     add_start_subcommand(sub_parsers)
     add_stop_subcommand(sub_parsers)
@@ -52,22 +50,6 @@ def parse_arguments():
 
     return parser.parse_args()
 
-def add_repair_status_subcommand(sub_parsers):
-    parser_repair_status = sub_parsers.add_parser("repair-status",
-                                                  description="Show status of all repairs and schedules")
-    parser_repair_status.add_argument("-k", "--keyspace", type=str,
-                                      help="Print status(es) for a specific keyspace")
-    parser_repair_status.add_argument("-t", "--table", type=str,
-                                      help="Print status(es) for a specific table (Must be specified with keyspace)")
-    parser_repair_status.add_argument("-u", "--url", type=str,
-                                      help="The host to connect to with the format (http://<host>:port)",
-                                      default=None)
-    parser_repair_status.add_argument("-i", "--id", type=str,
-                                      help="Print verbose status for a specific job")
-    parser_repair_status.add_argument("-l", "--limit", type=int,
-                                      help="Limit the number of tables or virtual nodes printed (-1 to disable)",
-                                      default=-1)
-    parser_repair_status.set_defaults(func=repair_status)
 
 def add_repairs_subcommand(sub_parsers):
     parser_repairs = sub_parsers.add_parser("repairs",
@@ -86,6 +68,7 @@ def add_repairs_subcommand(sub_parsers):
                                 default=-1)
     parser_repairs.add_argument("--hostid", type=str,
                                 help='Show repairs for the specified host id')
+
 
 def add_schedules_subcommand(sub_parsers):
     parser_schedules = sub_parsers.add_parser("schedules",
@@ -106,30 +89,6 @@ def add_schedules_subcommand(sub_parsers):
                                   help="Limit the number of tables or virtual nodes printed (-1 to disable)",
                                   default=-1)
 
-def add_repair_config_subcommand(sub_parsers):
-    parser_repair_config = sub_parsers.add_parser("repair-config",
-                                                  description="Show repair configuration")
-    parser_repair_config.add_argument("-k", "--keyspace", type=str,
-                                      help="Show config for a specific keyspace")
-    parser_repair_config.add_argument("-t", "--table", type=str,
-                                      help="Show config for a specific table")
-    parser_repair_config.add_argument("-i", "--id", type=str,
-                                      help="Show config for a specific job")
-    parser_repair_config.add_argument("-u", "--url", type=str,
-                                      help="The host to connect to with the format (http://<host>:port)",
-                                      default=None)
-
-def add_trigger_repair_subcommand(sub_parsers):
-    parser_trigger_repair = sub_parsers.add_parser("trigger-repair",
-                                                   description="Trigger a single repair")
-    parser_trigger_repair.add_argument("-u", "--url", type=str,
-                                       help="The host to connect to with the format (http://<host>:port)",
-                                       default=None)
-    required_args = parser_trigger_repair.add_argument_group("required arguments")
-    required_args.add_argument("-k", "--keyspace", type=str,
-                               help="Keyspace where the repair should be triggered", required=True)
-    required_args.add_argument("-t", "--table", type=str,
-                               help="Table where the repair should be triggered", required=True)
 
 def add_run_repair_subcommand(sub_parsers):
     parser_trigger_repair = sub_parsers.add_parser("run-repair",
@@ -145,6 +104,7 @@ def add_run_repair_subcommand(sub_parsers):
     required_args.add_argument("-t", "--table", type=str,
                                help="Table where the repair should be triggered", required=False)
 
+
 def add_start_subcommand(sub_parsers):
     parser_config = sub_parsers.add_parser("start",
                                            description="Start ecChronos service")
@@ -153,11 +113,13 @@ def add_start_subcommand(sub_parsers):
     parser_config.add_argument("-p", "--pidfile", type=str,
                                help="Pidfile where to store the pid, default $ECCHRONOS_HOME/ecc.pid")
 
+
 def add_stop_subcommand(sub_parsers):
     parser_stop = sub_parsers.add_parser("stop",
                                          description="Stop ecChronos service")
     parser_stop.add_argument("-p", "--pidfile", type=str,
                              help="Pidfile where to retrieve the pid, default $ECCHRONOS_HOME/ecc.pid")
+
 
 def add_status_subcommand(sub_parsers):
     parser_status = sub_parsers.add_parser("status",
@@ -166,36 +128,10 @@ def add_status_subcommand(sub_parsers):
                                help="The host to connect to with the format (http://<host>:port)",
                                default=None)
 
-def repair_status(arguments):
-    request = rest.RepairSchedulerRequest(base_url=arguments.url)
-
-    if arguments.id:
-        result = request.get(job_id=arguments.id)
-        if result.is_successful():
-            table_printer.print_verbose_repair_job(result.data, arguments.limit)
-        else:
-            print(result.format_exception())
-    elif arguments.table:
-        if not arguments.keyspace:
-            print("Must specify keyspace")
-            sys.exit(1)
-        result = request.list(keyspace=arguments.keyspace, table=arguments.table)
-        if result.is_successful():
-            table_printer.print_repair_jobs(result.data, arguments.limit)
-        else:
-            print(result.format_exception())
-    else:
-        result = request.list(keyspace=arguments.keyspace)
-        if result.is_successful():
-            table_printer.print_repair_jobs(result.data, arguments.limit)
-        else:
-            print(result.format_exception())
-
 
 def schedules(arguments):
     #pylint: disable=too-many-branches
     request = rest.V2RepairSchedulerRequest(base_url=arguments.url)
-    printer = table_printer_v2
     full = False
     if arguments.id:
         if arguments.full:
@@ -205,7 +141,7 @@ def schedules(arguments):
             result = request.get_schedule(job_id=arguments.id)
 
         if result.is_successful():
-            printer.print_schedule(result.data, arguments.limit, full)
+            table_printer.print_schedule(result.data, arguments.limit, full)
         else:
             print(result.format_exception())
     elif arguments.full:
@@ -217,24 +153,23 @@ def schedules(arguments):
             sys.exit(1)
         result = request.list_schedules(keyspace=arguments.keyspace, table=arguments.table)
         if result.is_successful():
-            printer.print_schedules(result.data, arguments.limit)
+            table_printer.print_schedules(result.data, arguments.limit)
         else:
             print(result.format_exception())
     else:
         result = request.list_schedules(keyspace=arguments.keyspace)
         if result.is_successful():
-            printer.print_schedules(result.data, arguments.limit)
+            table_printer.print_schedules(result.data, arguments.limit)
         else:
             print(result.format_exception())
 
 
 def repairs(arguments):
     request = rest.V2RepairSchedulerRequest(base_url=arguments.url)
-    printer = table_printer_v2
     if arguments.id:
         result = request.get_repair(job_id=arguments.id, host_id=arguments.hostid)
         if result.is_successful():
-            printer.print_repairs(result.data, arguments.limit)
+            table_printer.print_repairs(result.data, arguments.limit)
         else:
             print(result.format_exception())
     elif arguments.table:
@@ -243,52 +178,28 @@ def repairs(arguments):
             sys.exit(1)
         result = request.list_repairs(keyspace=arguments.keyspace, table=arguments.table, host_id=arguments.hostid)
         if result.is_successful():
-            printer.print_repairs(result.data, arguments.limit)
+            table_printer.print_repairs(result.data, arguments.limit)
         else:
             print(result.format_exception())
     else:
         result = request.list_repairs(keyspace=arguments.keyspace, host_id=arguments.hostid)
         if result.is_successful():
-            printer.print_repairs(result.data, arguments.limit)
+            table_printer.print_repairs(result.data, arguments.limit)
         else:
             print(result.format_exception())
 
 
-def repair_config(arguments):
-    request = rest.RepairConfigRequest(base_url=arguments.url)
-
-    if arguments.id:
-        if arguments.keyspace or arguments.table:
-            print("id must be specified alone")
-            sys.exit(1)
-        result = request.get(job_id=arguments.id)
-    else:
-        result = request.list(keyspace=arguments.keyspace, table=arguments.table)
-
-    if result.is_successful():
-        table_printer.print_table_config(result.data)
-    else:
-        print(result.format_exception())
-
 def run_repair(arguments):
     request = rest.V2RepairSchedulerRequest(base_url=arguments.url)
-    printer = table_printer_v2
     if not arguments.keyspace and arguments.table:
         print("--keyspace must be specified if table is specified")
         sys.exit(1)
     result = request.post(keyspace=arguments.keyspace, table=arguments.table, local=arguments.local)
     if result.is_successful():
-        printer.print_repairs(result.data)
+        table_printer.print_repairs(result.data)
     else:
         print(result.format_exception())
 
-def trigger_repair(arguments):
-    request = rest.RepairSchedulerRequest(base_url=arguments.url)
-    result = request.post(keyspace=arguments.keyspace, table=arguments.table)
-    if result.is_successful():
-        table_printer.print_repair_job(result.data)
-    else:
-        print(result.format_exception())
 
 def start(arguments):
     script_dir = os.path.dirname(os.path.realpath(__file__))
@@ -299,12 +210,14 @@ def start(arguments):
     command = "java {0} -cp {1} {2}".format(jvm_opts, class_path, SPRINGBOOT_MAIN_CLASS)
     run_ecc(ecchronos_home_dir, command, arguments)
 
+
 def get_class_path(conf_dir, ecchronos_home_dir):
     class_path = conf_dir
     jar_glob = os.path.join(ecchronos_home_dir, "lib", "*.jar")
     for jar_file in glob.glob(jar_glob):
         class_path += ":{0}".format(jar_file)
     return class_path
+
 
 def get_jvm_opts(conf_dir):
     jvm_opts = ""
@@ -313,6 +226,7 @@ def get_jvm_opts(conf_dir):
             if line.startswith("-"):
                 jvm_opts += "{0} ".format(line)
     return jvm_opts + "-Decchronos.config={0}".format(conf_dir)
+
 
 def run_ecc(cwd, command, arguments):
     if arguments.foreground:
@@ -334,6 +248,7 @@ def run_ecc(cwd, command, arguments):
             sys.stdout.write(line)
         proc.wait()
 
+
 def stop(arguments):
     script_dir = os.path.dirname(os.path.realpath(__file__))
     ecchronos_home_dir = os.path.join(script_dir, "..")
@@ -346,18 +261,8 @@ def stop(arguments):
         os.kill(pid, signal.SIGTERM)
     os.remove(pid_file)
 
-def status(arguments, print_running=False):
-    request = rest.RepairSchedulerRequest(base_url=arguments.url)
-    result = request.list()
-    print("Deprecated protocol, use 'repairs', 'schedules' or 'run-repair'")
-    if result.is_successful():
-        if print_running:
-            print("ecChronos is running")
-    else:
-        print("ecChronos is not running")
-        sys.exit(1)
 
-def status_v2(arguments, print_running=False):
+def status(arguments, print_running=False):
     request = rest.V2RepairSchedulerRequest(base_url=arguments.url)
     result = request.list_schedules()
     if result.is_successful():
@@ -367,24 +272,16 @@ def status_v2(arguments, print_running=False):
         print("ecChronos is not running")
         sys.exit(1)
 
+
 def run_subcommand(arguments):
-    if arguments.subcommand == "repair-status":
+    if arguments.subcommand == "repairs":
         status(arguments)
-        repair_status(arguments)
-    elif arguments.subcommand == "repairs":
-        status_v2(arguments)
         repairs(arguments)
     elif arguments.subcommand == "schedules":
-        status_v2(arguments)
+        status(arguments)
         schedules(arguments)
-    elif arguments.subcommand == "repair-config":
-        status(arguments)
-        repair_config(arguments)
-    elif arguments.subcommand == "trigger-repair":
-        status(arguments)
-        trigger_repair(arguments)
     elif arguments.subcommand == "run-repair":
-        status_v2(arguments)
+        status(arguments)
         run_repair(arguments)
     elif arguments.subcommand == "start":
         start(arguments)
@@ -392,6 +289,7 @@ def run_subcommand(arguments):
         stop(arguments)
     elif arguments.subcommand == "status":
         status(arguments, print_running=True)
+
 
 def main():
     run_subcommand(parse_arguments())
