+++ ecchronos-binary/src/pylib/ecchronoslib/table_printer.py
@@ -14,39 +14,39 @@
 #
 
 from __future__ import print_function
+from datetime import datetime
 from ecchronoslib import table_formatter
 
-# Disable duplicate code in deprecated class
-# Class will be removed
-# pylint: skip-file
 
-def print_verbose_repair_job(repair_job, max_lines):
-    if not repair_job.is_valid():
-        print('Repair job not found')
+def print_schedule(schedule, max_lines, full=False):
+    if not schedule.is_valid():
+        print('Schedule not found')
         return
 
     verbose_print_format = "{0:15s}: {1}"
 
-    print(verbose_print_format.format("Id", repair_job.job_id))
-    print(verbose_print_format.format("Keyspace", repair_job.keyspace))
-    print(verbose_print_format.format("Table", repair_job.table))
-    print(verbose_print_format.format("Status", repair_job.status))
-    print(verbose_print_format.format("Repaired(%)", repair_job.get_repair_percentage()))
-    print(verbose_print_format.format("Completed at", repair_job.get_last_repaired_at()))
-    print(verbose_print_format.format("Next repair", repair_job.get_next_repair()))
-    print(verbose_print_format.format("Recurring", repair_job.recurring))
+    print(verbose_print_format.format("Id", schedule.job_id))
+    print(verbose_print_format.format("Keyspace", schedule.keyspace))
+    print(verbose_print_format.format("Table", schedule.table))
+    print(verbose_print_format.format("Status", schedule.status))
+    print(verbose_print_format.format("Repaired(%)", schedule.get_repair_percentage()))
+    print(verbose_print_format.format("Completed at", schedule.get_last_repaired_at()))
+    print(verbose_print_format.format("Next repair", schedule.get_next_repair()))
+    print(verbose_print_format.format("Config", schedule.get_config()))
 
-    vnode_state_table = [["Start token", "End token", "Replicas", "Repaired at", "Repaired"]]
+    if full:
+        vnode_state_table = [["Start token", "End token", "Replicas", "Repaired at", "Repaired"]]
 
-    sorted_vnode_states = sorted(repair_job.vnode_states, key=lambda vnode: vnode.last_repaired_at_in_ms, reverse=True)
+        sorted_vnode_states = sorted(schedule.vnode_states, key=lambda vnode: vnode.last_repaired_at_in_ms,
+                                     reverse=True)
 
-    if max_lines > -1:
-        sorted_vnode_states = sorted_vnode_states[:max_lines]
+        if max_lines > -1:
+            sorted_vnode_states = sorted_vnode_states[:max_lines]
 
-    for vnode_state in sorted_vnode_states:
-        _add_vnode_state_to_table(vnode_state, vnode_state_table)
+        for vnode_state in sorted_vnode_states:
+            _add_vnode_state_to_table(vnode_state, vnode_state_table)
 
-    table_formatter.format_table(vnode_state_table)
+        table_formatter.format_table(vnode_state_table)
 
 def _add_vnode_state_to_table(vnode_state, table):
     entry = [vnode_state.start_token, vnode_state.end_token, ', '.join(vnode_state.replicas),
@@ -54,8 +54,17 @@ def _add_vnode_state_to_table(vnode_state, table):
 
     table.append(entry)
 
-def print_summary(repair_jobs):
-    status_list = [job.status for job in repair_jobs]
+def print_summary(schedules):
+    status_list = [schedule.status for schedule in schedules]
+    summary_format = "Summary: {0} completed, {1} on time, {2} blocked, {3} late, {4} overdue"
+    print(summary_format.format(status_list.count('COMPLETED'),
+                                status_list.count('ON_TIME'),
+                                status_list.count('BLOCKED'),
+                                status_list.count('LATE'),
+                                status_list.count('OVERDUE')))
+
+def print_repair_summary(repairs):
+    status_list = [repair.status for repair in repairs]
     summary_format = "Summary: {0} completed, {1} in queue, {2} blocked, {3} warning, {4} error"
     print(summary_format.format(status_list.count('COMPLETED'),
                                 status_list.count('IN_QUEUE'),
@@ -63,48 +72,56 @@ def print_summary(repair_jobs):
                                 status_list.count('WARNING'),
                                 status_list.count('ERROR')))
 
-def print_repair_jobs(repair_jobs, max_lines):
-    repair_jobs_table = [["Id", "Keyspace", "Table", "Status", "Repaired(%)",
-                          "Completed at", "Next repair", "Recurring"]]
-    sorted_repair_jobs = sorted(repair_jobs, key=lambda job: job.last_repaired_at_in_ms, reverse=True)
 
+def print_schedules(schedules, max_lines):
+
+    schedule_table = [["Id", "Keyspace", "Table", "Status", "Repaired(%)",
+                       "Completed at", "Next repair"]]
+    print("Snapshot as of", datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
+    print_schedule_table(schedule_table, schedules, max_lines)
+    print_summary(schedules)
+
+def print_repairs(repairs, max_lines=-1):
+    repair_table = [["Id", "Host Id", "Keyspace", "Table", "Status", "Repaired(%)",
+                     "Completed at"]]
+    print_repair_table(repair_table, repairs, max_lines)
+    print_repair_summary(repairs)
+
+def print_schedule_table(schedule_table, schedules, max_lines):
+
+    sorted_schedules = sorted(schedules, key=lambda x: (x.last_repaired_at_in_ms, x.repaired_ratio),
+                              reverse=False)
     if max_lines > -1:
-        sorted_repair_jobs = sorted_repair_jobs[:max_lines]
+        sorted_schedules = sorted_schedules[:max_lines]
 
-    for repair_job in sorted_repair_jobs:
-        repair_jobs_table.append(_convert_repair_job(repair_job))
-    table_formatter.format_table(repair_jobs_table)
+    for schedule in sorted_schedules:
+        schedule_table.append(_convert_schedule(schedule))
+    table_formatter.format_table(schedule_table)
 
-    print_summary(repair_jobs)
+def print_repair_table(repair_table, repairs, max_lines):
 
-def print_repair_job(repair_job):
-    repair_jobs_table = [["Id", "Keyspace", "Table", "Status", "Repaired(%)",
-                          "Completed at", "Next repair", "Recurring"]]
-    repair_jobs_table.append(_convert_repair_job(repair_job))
-    table_formatter.format_table(repair_jobs_table)
+    sorted_repairs = sorted(repairs, key=lambda x: (x.completed_at, x.repaired_ratio), reverse=False)
+    if max_lines > -1:
+        sorted_repairs = sorted_repairs[:max_lines]
 
-def _convert_repair_job(repair_job):
-    entry = [repair_job.job_id, repair_job.keyspace, repair_job.table, repair_job.status,
-             repair_job.get_repair_percentage(), repair_job.get_last_repaired_at(), repair_job.get_next_repair(),
-             repair_job.recurring]
+    for repair in sorted_repairs:
+        repair_table.append(_convert_repair(repair))
+    table_formatter.format_table(repair_table)
 
+def print_repair(repair):
+    repair_table = [["Id", "Host Id", "Keyspace", "Table", "Status", "Repaired(%)",
+                     "Completed at"]]
+    repair_table.append(_convert_repair(repair))
+    table_formatter.format_table(repair_table)
+
+
+def _convert_repair(repair):
+    entry = [repair.job_id, repair.host_id, repair.keyspace, repair.table, repair.status,
+             repair.get_repair_percentage(), repair.get_completed_at()]
     return entry
 
-def print_table_config(config_data):
-    config_table = [["Id", "Keyspace", "Table", "Interval",
-                     "Parallelism", "Unwind ratio", "Warning time", "Error time"]]
-    if isinstance(config_data, list):
-        sorted_config_data = sorted(config_data, key=lambda config: (config.keyspace, config.table))
-        for config in sorted_config_data:
-            if config.is_valid():
-                config_table.append(_convert_config(config))
-    elif config_data.is_valid():
-        config_table.append(_convert_config(config_data))
-
-    table_formatter.format_table(config_table)
-
-def _convert_config(config):
-    entry = [config.job_id, config.keyspace, config.table, config.get_repair_interval(), config.repair_parallelism,
-             config.repair_unwind_ratio, config.get_repair_warning_time(), config.get_repair_error_time()]
+def _convert_schedule(schedule):
+    entry = [schedule.job_id, schedule.keyspace, schedule.table, schedule.status,
+             schedule.get_repair_percentage(), schedule.get_last_repaired_at(), schedule.get_next_repair()]
 
     return entry
