+++ ecchronos-binary/src/pylib/ecchronoslib/types.py
@@ -46,35 +46,6 @@ class VnodeState(object):
         return datetime.datetime.fromtimestamp(self.last_repaired_at_in_ms / 1000).strftime('%Y-%m-%d %H:%M:%S')
 
 
-class RepairJob(object):
-    # pylint: disable=too-many-instance-attributes
-
-    def __init__(self, data):
-        self.keyspace = data["keyspace"] if "keyspace" in data else "<UNKNOWN>"
-        self.table = data["table"] if "table" in data else "<UNKNOWN>"
-        self.last_repaired_at_in_ms = int(data["lastRepairedAtInMs"] if "lastRepairedAtInMs" in data else -1)
-        self.repaired_ratio = float(data["repairedRatio"] if "repairedRatio" in data else 0)
-        self.status = data["status"] if "status" in data else "<UNKNOWN>"
-        self.next_repair_in_ms = int(data["nextRepairInMs"] if "nextRepairInMs" in data else -1)
-        self.recurring = data["recurring"] if "recurring" in data else "<UNKNOWN>"
-        self.job_id = data["id"] if "id" in data else "<UNKNOWN>"
-
-    def is_valid(self):
-        return self.keyspace != "<UNKNOWN>"
-
-    def get_last_repaired_at(self):
-        if self.last_repaired_at_in_ms == -1:
-            return "-"
-        return datetime.datetime.fromtimestamp(self.last_repaired_at_in_ms / 1000).strftime('%Y-%m-%d %H:%M:%S')
-
-    def get_repair_percentage(self):
-        return "{0:.2f}".format(self.repaired_ratio * 100.0)
-
-    def get_next_repair(self):
-        if self.next_repair_in_ms == -1:
-            return "-"
-        return datetime.datetime.fromtimestamp(self.next_repair_in_ms / 1000).strftime('%Y-%m-%d %H:%M:%S')
-
 class Job(object):
 
     def __init__(self, data):
@@ -89,6 +60,7 @@ class Job(object):
 
     def get_repair_percentage(self):
         return "{0:.2f}".format(self.repaired_ratio * 100.0)
+
 
 class Repair(Job):
 
@@ -133,37 +105,3 @@ class FullSchedule(Schedule):
         if "virtualNodeStates" in data:
             for vnode_data in data["virtualNodeStates"]:
                 self.vnode_states.append(VnodeState(vnode_data))
-
-class VerboseRepairJob(RepairJob):
-    def __init__(self, data):
-        RepairJob.__init__(self, data)
-        self.vnode_states = []
-        if "virtualNodeStates" in data:
-            for vnode_data in data["virtualNodeStates"]:
-                self.vnode_states.append(VnodeState(vnode_data))
-
-
-class TableConfig(object):
-    # pylint: disable=too-many-instance-attributes
-
-    def __init__(self, data):
-        self.job_id = data["id"] if "id" in data else "<UNKNOWN>"
-        self.keyspace = data["keyspace"] if "keyspace" in data else "<UNKNOWN>"
-        self.table = data["table"] if "table" in data else "<UNKNOWN>"
-        self.repair_interval_in_ms = int(data["repairIntervalInMs"] if "repairIntervalInMs" in data else 0)
-        self.repair_parallelism = data["repairParallelism"] if "repairParallelism" in data else "<UNKNOWN>"
-        self.repair_unwind_ratio = float(data["repairUnwindRatio"] if "repairUnwindRatio" in data else 0)
-        self.repair_warning_time_in_ms = int(data["repairWarningTimeInMs"] if "repairWarningTimeInMs" in data else 0)
-        self.repair_error_time_in_ms = int(data["repairErrorTimeInMs"] if "repairErrorTimeInMs" in data else 0)
-
-    def is_valid(self):
-        return self.keyspace != "<UNKNOWN>"
-
-    def get_repair_interval(self):
-        return parse_interval(self.repair_interval_in_ms)
-
-    def get_repair_warning_time(self):
-        return parse_interval(self.repair_warning_time_in_ms)
-
-    def get_repair_error_time(self):
-        return parse_interval(self.repair_error_time_in_ms)
