+++ ecchronos-binary/src/test/behave/features/ecctool-get-repairs.feature
@@ -13,7 +13,7 @@ Feature: ecctool repairs
     And the output should contain a repair row for test2.table1
     And the output should contain a repair row for test2.table2
     And the output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: List repairs with a limit
     Given we have access to ecctool
@@ -21,7 +21,7 @@ Feature: ecctool repairs
     Then the output should contain a valid repair header
     And the output should contain 1 repair rows
     And the output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: List repairs for keyspace test2
     Given we have access to ecctool
@@ -30,7 +30,7 @@ Feature: ecctool repairs
     And the output should contain a repair row for test2.table1
     And the output should contain a repair row for test2.table2
     And the output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: List repairs for keyspace test2 with a limit
     Given we have access to ecctool
@@ -38,7 +38,7 @@ Feature: ecctool repairs
     Then the output should contain a valid repair header
     And the repair output should contain a valid repair row for test2..*
     And the output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: List the repair test2.table2
     Given we have access to ecctool
@@ -46,14 +46,14 @@ Feature: ecctool repairs
     Then the output should contain a valid repair header
     And the output should contain a repair row for test2.table2
     And the output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: List the repairs test2.table2 with a limit
     Given we have access to ecctool
     When we list repairs test2.table2 with a limit of 15
     Then the output should contain a repair row for test2.table2
     And the output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: List the repair test2.table2 with hostid
     Given we have access to ecctool
@@ -61,4 +61,4 @@ Feature: ecctool repairs
     Then the output should contain a valid repair header
     And the output should contain a repair row for test2.table2
     And the output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
