+++ ecchronos-binary/src/test/behave/features/ecctool-run-repair.feature
@@ -6,7 +6,7 @@ Feature: ecctool run-repair
     Then the repair output should contain a valid header
     And the repair output should contain a valid repair row for test2.table2
     And the repair output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: Run local repair for keyspace test2
     Given we have access to ecctool
@@ -15,7 +15,7 @@ Feature: ecctool run-repair
     And the repair output should contain a valid repair row for test2.table2
     And the repair output should contain a valid repair row for test2.table1
     And the repair output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: Run local repair for all keyspaces and tables
     Given we have access to ecctool
@@ -26,7 +26,7 @@ Feature: ecctool run-repair
     And the repair output should contain a valid repair row for test.table2
     And the repair output should contain a valid repair row for test.table1
     And the repair output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: Run cluster-wide repair for keyspace test2 and table table2
     Given we have access to ecctool
@@ -35,7 +35,7 @@ Feature: ecctool run-repair
     And the repair output should contain a valid repair row for test2.table2
     And the repair output should contain a valid repair row for test2.table2
     And the repair output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: Run cluster-wide repair for keyspace test2
     Given we have access to ecctool
@@ -46,7 +46,7 @@ Feature: ecctool run-repair
     And the repair output should contain a valid repair row for test2.table1
     And the repair output should contain a valid repair row for test2.table1
     And the repair output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
 
   Scenario: Run cluster-wide repair for all keyspaces and tables
     Given we have access to ecctool
@@ -61,4 +61,4 @@ Feature: ecctool run-repair
     And the repair output should contain a valid repair row for test2.table1
     And the repair output should contain a valid repair row for test2.table1
     And the repair output should not contain more rows
-    And the output should contain summary
+    And the output should contain a valid repair summary
